refac: reorganize executor files
